Upgrade to Qt 6 (#130)
Why? Open Source Qt 5 will stop receiving updates rather sooner than later so it's worth switching to Qt 6 to remain compatible with modern systems.

Reviewed-on: http://vub63vv26q6v27xzv2dtcd25xumubshogm67yrpaz2rculqxs7jlfqad.onion/torzu-emu/torzu/pulls/130

+++ src/shader_recompiler/frontend/ir/microinstruction.cpp
@@ -250,7 +250,8 @@ bool Inst::AreAllArgsImmediates() const {
     if (op == Opcode::Phi) {
         throw LogicError("Testing for all arguments are immediates on phi instruction");
     }
-    return std::all_of(args.begin(), args.begin() + NumArgs(),
+    return std::all_of(args.begin(),
+                       args.begin() + static_cast<ptrdiff_t>(NumArgs()),
                        [](const IR::Value& value) { return value.IsImmediate(); });
 }
